Move direct tool execute route test
